Attempt at fixing installation of Python if there are spaces in the paths
A lot of these commands would have their parameters broken up in that case.
I'm not sure if the variable setting would work correctly then. The bindir is the one command that I didn't add quotes to, because that would cause more quotes to show up when the bindir is combined with new paths.

Attempt at fixing Ultimaker/Cura#7580.