Update urllib3 to version 1.26.2
This fixes security issue CVE-2020-26137, an attack where if the user can control the HTTP method, they can control the entire request (such as the URL it's requesting). It doesn't affect Cura since we hard-code the request method to PUT or GET in our code, and I don't think any plug-in would allow the user to change the request method either. But you never know.
It was reported to also fix security issue CVE-2020-7212, but we were on a too old version to be affected by that.

Contributes to issue CURA-8400.